requirements.txt
Django==4.2.20
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pytest==7.1.3
pytest==8.3.5
djoser 